Query block sizes at runtime
We now calculate block sizes while the simulation is ongoing, rather
than after the simulation is done. We do this because we want to be able
to prune the ledgers in the orderer and peer, so that the simulation's
disk requirements can be contained.

This changeset also adds logging levels, and gates debugging statements
behind the correspoding level. We didn't split this into a separate
changeset because we're bored.

Signed-off-by: Kostas Christidis <kostas@christidis.io>